+++ CHANGELOG.rst
@@ -17,6 +17,7 @@ Added
     - Support for multiple ontology links in ``_get_ontology_link()`` (#431)
     - Hack for providing correct HPO ontology into links (#431)
     - Rendering for HPO term links (#431)
+    - Rendering for performer and perform date (#187)
 
 Changed
 -------